Code for Beginners on creating a management website
This website is created using HTML, PHP and MYSQL has its backend server to collect the costumer's information. First the home is placed. Then comes the registration form which basically asks the costumer about their personal details and the item they wanted to purchase. Once all the informations are filled the details are successfullly registered and the information are stored in the DB. Then comes the login and logout sessions provided it can be logged in only with an id and password.